requirements.txt
streamlit>=1.30.0
streamlit-folium>=0.18.0
folium>=0.15.0
pandas>=2.0.0
icalendar>=5.0.0
requests>=2.31.0
pydantic>=2.5.0
python-dateutil>=2.8.0
pytz>=2024.1
rapidfuzz>=3.0.0
python-dotenv>=1.0.0
google-genai>=0.1.0
googlemaps>=4.10.0
tavily-python>=0.3.0
